Update to the release service interface for file upload, should be less error prone this way since the github api returns an almost complete url.
Create a whole slew of tasks. one for each asset to upload per variant as well as one task for each variant that depends on all of the other upload tasks. Should help simplify hooking this into your own build scripts. A few tweaks to the test. Still not the best but better than nothing.